reorganize clusterctl code into separate files
